requirements.txt
discord.py
python-dotenv
firebase_admin
apscheduler